Ajustes para deixar o cadastro e pesquisa de clientes e equipamentos funcionando 100%. Inclusão de filtro na págiina de listagem de equipamentos. Inclusão de arquivo com as funcionalidades de paginação, PageUtils. Inclusão dos menus com os icones corretos para as páginas de cadastro e pesquisa de clientes e equipamentos. AAjustes dos relacionamentos de equipamentos com o cliente para permitir a listagem e pesquisa de equipamentos por cliente.

+++ cmkservicos/src/main/java/br/com/iridiumit/cmkservicos/repository/Clientes.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.iridiumit.cmkservicos.modelos.Cliente;
 
-public interface Clientes extends JpaRepository<Cliente, Long>{
+public interface Clientes extends JpaRepository<Cliente, Integer>{
 
 	List<Cliente> findByNomeContainingIgnoreCaseAndAtivo(String nome, boolean ativo);
